Fixed bug where two players couldn't place ships
I guess Python links objects by reference, so I had one dict that contained all the ships, and passed it to the player object on initialization. I expected then that each player would have their own, but identical, list of ships. This was not the case. So I created two separate but identical dicts first and passed those to each individual player.